requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking
#
aiohttp==3.8.4
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
cffi==1.15.1
    # via pynacl
charset-normalizer==3.1.0
    # via aiohttp
discord-py[voice]==2.3.0
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
pycparser==2.21
    # via cffi
pynacl==1.5.0
    # via discord-py
python-dotenv==1.0.0
    # via -r requirements.in
yarl==1.9.2
    # via aiohttp
